Add ability to work as /usr/bin/git on MSYS2
This allows for a user to link /mingwXX/bin/git to /usr/bin/git and use it
as their MSYS2 install's git.  This change is needed so that lookups using
system_path work correctly in this case.

Note that if the user does this, they will also need to modify
/mingwXX/etc/gitconfig so that http.sslCAInfo points to
/mingwXX/ssl/certs/ca-bundle.crt or similar.