winansi.c: Fix colourization on Cygwin pseudo terminals.
Git only colours the output and uses pagination if isatty() returns 1.  MSys and Cygwin emulate pseudo terminals via named pipes, meaning that isatty() returns 0.

Commit 3adef8de5513390459e1caa999e7774adb89a44d fixed this for MSys terminals, but not Cygwin.

The named pipes that Cygwin and Msys use are very similar.  MSys PTY pipes are called 'msys-*-pty*' and Cygwin uses 'cygwin-*-pty*'.  This commit modifies the existing check to allow both MSys and Cygwin PTY pipes to be identified as TTYs.

Note that Pagination is still broken on Cygwin.  less.exe is spawned (as seen in Process Explorer and using GIT_TRACE=1), but the output is not being piped into it.

This partially fixes https://github.com/git-for-windows/git/issues/267