mingw: let `core.fsyncObjectFiles` default to the batch mode
For several years, Git for Windows ran with the `core.fsyncObjectFiles`
mode enabled. This forced the disk to be synchronized after each and
every loose object file was written, adding a substantial layer of
safety (otherwise, a crash of the Operating System might have left files
in place of the correct size, but filled with NUL bytes).

In contrast to what this maintainers' assessment of the performance
implications was (namely, that they should be negligible), it turns out
that in particular when working on large monorepos, it _does_ have a
high impact.

In late 2021, Git learned a batched mode in the meantime, a mode that
would batch those synchronizations, which strikes a better balance
between safety and performance. Let's switch to that mode as the new
default.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>