crtdbg: fix MSVC CRTDBG memory leak reporting
Teach git to report memory leaks both to a file and to the
VS debug output window.

From the SDK command line, build Git as follows:

	make MSVC=1 DEBUG=1 CRTDBG=1

When built this way, leaks will always be reported to the
Visual Studio debug output window.

Enable GIT_TRACE_CRTDBG and the leaks will also be reported
to the trace target.  This works from both the command line
and from Visual Studio.

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>