mingw: take argv[0] from command line parameters
Instead of basing argv[0] on the executable path, use argv[0] as
provided on the command line.

Setting argv[0] to git-<builtin> is currently the only way to prevent
git from falling back to calling a helper executable. This change will
allow the git-wrapper executable, that Git for Windows uses, to signal
to git to handle the command as a builtin.

Related to https://github.com/git-for-windows/git/issues/1496

Signed-off-by: Kim Gybels <kgybels@infogroep.be>