pass extended startup info directly
The extended startup info is not passed to CreateProcess. It is passed
indirectly as &si.StartupInfo. Since StartupInfo is the first member of
the extended startup info struct, the address of si.StartupInfo
coincides with the address of si.

Although it makes no difference functionally, it seems cleaner to pass
extended startup info directly, as suggested also by the
UpdateProcThreadAttribute example:

https://docs.microsoft.com/en-us/windows/desktop/api/processthreadsapi/nf-processthreadsapi-updateprocthreadattribute

Signed-off-by: Clemens Buchacher <drizzd@gmx.net>