doc: remove unneeded TODO for release_commit_memory
The code that was merged with commit 1102405 left in an TODO that
is no longer relevant. It seems as if we can remove this todo.
util seems to not be a field of the struct commit,
definition of struct commit --> commit.c:27
The commit list also, does not contain a field for util.

Signed-off-by: Albert Burt <aburthinds@gmail.com>