Visual Studio solution: check for updates of vcpkg
The Microsoft vcpkg package manager recieves continual updates. The
pre-built vs/master git.sln Visual Studio solution would clone and
compile the vcpkg dependencies if missing, but did not check for
updates. Fix this user-land issue.

The Azure Continuous Integration (CI) system would pre-build the vcpkg
files and expand them in place rather than clone and rebuild them.

Unfortunately the CI logic and the update logic are in conflict as both
detect/use the presence/absence of the same files via the git.sln
Prebuild Event.

The .git directory is not present in the Azure-CI, which conflicts
with using a `git pull` update..

Add a sentinel file to the Azure-CI to indicate that the CI is in use,
and use that as the git.sln conditional PrebuildEvent for checking for
vcpkg updates in user-land.

Update the contrib/buildsystems/Generators/Vcxproj.pm.

Remove the condition on 'Pre-Build' event for installation, and thus
always check for updates if the vcpkg manager is already present.

The `vcpkg_install.bat' file name is retained to avoid churn. The batch
file now also handles the vcpkg update check.

See https://github.com/microsoft/vcpkg/issues/9148 for update documentation,
however the use of "./vcpkg" is a bashism, not suited to a cmd.bat file.
Use `vcpkg`, no need for the .exe part, removing the `./` should be
sufficient.

Signed-off-by: Philip Oakley <philipoakley@iee.email>
---
corrected some bad logic and confusion that success is zero
though logical true for the && cases, and the reverse
sequence of existance checks.

still contains residual artefacts for dev assistance.