compat/mingw: avoid implict declaration of gmtime_r, localtime_r
Commit 3ecd153a3b ("compat/mingw: support MSys2-based MinGW build",
2016-01-14) fixed a range of Windows build errors. However some compile
warnings were left for a future patch.

Now that all warnings being treated as errors, the conditional failure
to declare gmtime_r() and localtime_r() in mingw.h is a 'make' compilation
failure. This is the patch to fix such warnings (treated as erors).

Avoid implicit declaration by removing the inverted ifndef conditional.

Signed-off-by: Philip Oakley <philipoakley@iee.email>