init: introduce advice.defaultBranchName
This introduces a way to silence that warning upon `git init` that says
that the default initial branch name is subject to change, without
explicitly overriding Git's choice.

This comes in handy in Git for Windows, where late in September 2020 an
installer option was introduced to either specify the default branch
name or just go with whatever Git chooses.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>