fsmonitor: handle RegisterApplicationRestart() and CTRL_C_EVENT
Teach the fsmonitor--daemon to register itself with Resource Manager
so that the Git for Windows installer can stop and restart them during
an upgrade.  Without this, the installer cannot update the .EXE and
.DLLs that the daemon process is using.

Also teach the fsmonitor--daemon to handle CTRL-C events properly.
The Resource Manager sends a CTRL_C_EVENT to the process to ask it to
shutdown during the upgrade.  (The system also sends this event when
the user types a CTRL-C in the console window.)

NEEDSWORK: The solution here substitutes a trivial `git version`
command instead of the proper command to restart the fsmonitor daemon
because of Windows complications.  We don't really care if the daemon
restarts immediately after the installer finishes or if it is started
later by the first client command.  Our primary goal here is to let
the installer stop running daemons and do the update.

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>