Allow overriding which POSIX shell to use
Right now, the compiled-in default is used always, which can be
overridden at compile time via the `SHELL_PATH` variable.

However, in Git for Windows' context, there is a scenario where this is
not good enough: the BusyBox flavor of MinGit wants to use BusyBox'
`ash` instead.

To allow for that, let's introduce a new config variable: `core.shell`.
Its value is expected to be the absolute path to a valid, working POSIX
shell. This shell will be then used by `git.exe` whenever it needs to
execute something via shell, such as moderately complex aliases.

And it is not only Git that is expected to use that `ash`, it is also
OpenSSH (e.g. when running any configured `ProxyCommand`): programs
spawned from `git.exe` may need to run a shell and look at the
environment variable `SHELL` for that. Therefore, let's set that, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>