merge: add new --into option
This adds the ability to perform merges in a bare repository, or into
a branch that is not checked out.  However, it only computes the
toplevel tree, it does not create a commit object or update any refs.
Thus, it would probably only be used as a step in a process, like this:

   NEWTREE=$(git merge --into $PARENT1 $PARENT2)
   test $? -eq 0 || die There were conflicts...
   NEWCOMMIT=$(git commit-tree $NEWTREE -p $PARENT1 $PARENT2)
   git update-ref $PARENT1 $NEWCOMMIT

Since this is so different that what builtin/merge.c does, I'm not even
sure if it makes sense to add here.  Perhaps it belongs in
builtin/merge-recursive.c (though there's also some weirdness there) or
as a new toplevel command.  Some things to note about this new --into
mode:

It's incompatible with most existing flags to merge:
    * Anything dealing with a commit message:
      * -m
      * --[no-]edit
      * --cleanup
      * --log
      * --signoff
      * --[no-]gpg-sign/-S
      * --file/-F
    * Anything dealing with commits:
      * --[no-]commit
      * --squash
    * Anything dealing with the working tree:
      * --autostash
      * --overwrite-ignore
    * Anything dealing with the index
      * --[no-]rerere-autoupdate
    * Anything dealing with extraneous output:
      * --stat/--summary
      * --progress
      * --verbose
      * --quiet

It implies a few other options:
    * --no-verify (no commit will be created, also we have no worktree)
    * --no-ff
    * --no-progress (progress would destroy scriptable output)

It only allows a few existing options:
    * --verify-signatures
    * --allow-unrelated-histories
    * --strategy-option/-X

It adds a few new options, specific to this mode:
    * --output (a file for messages about conflicts and three-way merged
      files to be written to)
    * --unmerged (a file to record unmerged filenames, one per line.
      Perhaps these should be NUL-delimited instead of newline
      delimited.  *shrug*)

Signed-off-by: Elijah Newren <newren@gmail.com>