Use env as interpreter for all exe and scripts.
Git requires a MSYS2 perl. However, there are many distributions
of perl. If another perl is in the PATH in front of the MSYS2 one,
git will mistakenly find it. The shebang is right #!/usr/bin/perl,
but actually git only recognizes perl and finds the first one in
the PATH.

This pull request removes the naïve parser of shebang, and just finds
`env.exe` as the interpreter. It should be in the `PATH` in MSYS2.

Signed-off-by: Yuyi Wang <Strawberry_Str@hotmail.com>