add helper funcs for sync with wsl file mode bits
In WSL(2), we can set the mode for an NTFS file, for example:

$ chmod 0755 /mnt/d/test/a.sh

provided that WSL has enabled metadata mounting NTFS volumes.

In order to facilitate better collaboration between the Windows
version of Git and the WSL version of Git, we can make the Windows
version of Git also support reading and writing NTFS file modes
in a manner compatible with WSL.

In this commit, I've added two helper functions that read and write
file mode bits in a manner compatible with WSL.

There are several ways to enable metadata loading for NTFS volumes
in WSL, one of which is to modify /etc/wsl.conf by adding:

```
[automount]
enabled = true
options = "metadata,umask=027,fmask=117"
```

And reboot.

It's important to note that this modification is compatible with,
but does not depend on WSL. The helper functions in this patch
can operate independently and function normally on devices
where WSL is not installed or properly configured.

Signed-off-by: xungeng <xungeng@gmail.com>