sync file mode bits with WSL
In this commit, I use the previous added helper functions to
finally make the current version of git compatible with WSL git on
file mode bits.

With this commit, git can checkout files with x mode set, and
won't report modifications if the repo is checkout by wsl git
and there are some files with x mode set.

Signed-off-by: xungeng <xungeng@gmail.com>