macos-installer/Makefile: debug dump linkage of build targets
Run `otool -L` on each of the major executables and print the name of
the CPU architecture(s) and the name and version of the referenced
dynamic libraries.

In a later commit, we might want to verify that certain minimim
versions were linked or that there are no unexpected third-party
libraries in the linkage (such as an unexpected homebrew).

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>